requirements.txt
fastapi
uvicorn
openai
python-dotenv
faiss-cpu
langchain
langchain-openai
tiktoken
pydantic